feat(creation-service): Add interfaces for controller and pdf
