+++ services/creation-service/main.go
@@ -5,7 +5,8 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/hingew/hsfl-master-ai-cloud-engineering/creation-service/api/handler"
+	"github.com/hingew/hsfl-master-ai-cloud-engineering/creation-service/pdf"
+	"github.com/hingew/hsfl-master-ai-cloud-engineering/creation-service/pdf/controller"
 	"github.com/hingew/hsfl-master-ai-cloud-engineering/lib/config"
 	"github.com/hingew/hsfl-master-ai-cloud-engineering/lib/router"
 )
@@ -15,19 +16,21 @@ type ApplicationConfig struct {
 }
 
 func main() {
-    var appConfig *ApplicationConfig
+	var appConfig *ApplicationConfig
 
-    if err := config.Load(appConfig); err != nil {
+	if err := config.Load(appConfig); err != nil {
 		log.Fatalf("could not load application configuration: %s", err.Error())
-    }
+	}
 
-    router := router.New()
-    router.POST("/create", handler.Create())
+	pdf := pdf.New()
+	controller := controller.NewController(pdf)
+
+	router := router.New()
+	router.POST("/create", controller.CreatePdf)
 
 	addr := fmt.Sprintf("0.0.0.0:%d", *&appConfig.Port)
 	if err := http.ListenAndServe(addr, router); err != nil {
 		log.Fatalf("error while listen and serve: %s", err.Error())
 	}
-
 
 }
