+++ services/creation-service/pdf/controller/controller_interface.go
@@ -1,0 +1,8 @@
+package controller
+
+import "net/http"
+
+type ControllerInterface interface {
+	CreatePdf(http.ResponseWriter, *http.Request)
+}
+
