+++ services/creation-service/pdf/pdf_interface.go
@@ -1,0 +1,8 @@
+package pdf
+
+import "bytes"
+
+type Pdf interface {
+	AddElement()
+	Out() (*bytes.Buffer, error)
+}
